backend/requirements.txt
fastapi
uvicorn
hyperliquid-python-sdk
python-dotenv
pydantic
sqlmodel
eth-account
aiohttp>=3.9.0
loguru
websockets<14.0
sqlalchemy-libsql
numpy
pandas
feedparser
httpx
pyyaml>=6.0
python-telegram-bot>=20.0
aiosqlite>=0.20.0
